OGC Refactoring Part 7 - Move profile specific research queries to package /ogc/profiles
